backend/requirements.txt
fastapi
uvicorn
websockets
sqlalchemy
pydantic
openai
python-dotenv
cryptography